+++ format.py
@@ -1,0 +1,20 @@
+# Josh Lim
+# Comp Sci 30 P4
+# 12/09/2019
+# File for formating text in the game
+
+
+# Code adapted from
+# Stack Overflow: how to print colored text in terminal in python
+# https://stackoverflow.com/questions/287871/
+# how-to-print-colored-text-in-terminal-in-python
+class Colors:
+    """Classes for adding color to the python terminal"""
+    HEADER = '\033[95m'
+    BLUE = '\033[94m'
+    GREEN = '\033[92m'
+    YELLOW = '\033[93m'
+    RED = '\033[91m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
+    END = '\033[0m'