+++ locations.py
@@ -2,22 +2,3 @@
 # Comp Sci 30 P4
 # 11/07/2019
 # Dictionary and list containing all the locations
-
-
-escapeRoom = {
-    "description": "you find yourself in a room, in it you see..",
-    "cabinet": ["go back"],
-    "chest": ["go back"],
-    "bookcase": ["go back"],
-    "painting": ["go back"],
-    "desk": ["go back"],
-    "shelve": ["go back"],
-    "table": ["go back"],
-    "chair": ["go back"],
-}
-
-cabinet = {
-    "description": "",
-    "key": True,
-    "description": "",
-}