+++ menu.py
@@ -54,3 +54,19 @@ def menuLDesc(menuName, key):
     for i in menuName[key]:  # checks which path the player chose
         if option == "go back":  # if go back, reruns previous menu
             choosenPath(menuName)
+
+
+def printNDict(dictName):
+    """Prints out content in a nested dictionary with up to 2 layers,
+    takes dictName as the dictionary name.
+    """
+    for dictName, key in dictName.items():
+        """loops through the keys in the nested dictionary, dictName, and then
+        prints the nested dictionary names along with its values.
+        """
+        print(dictName.title() + ":")  # prints nested dictionary name
+        for items in key:
+            """loops through nested dictionary key values and and prints the
+            key along with the value with a tab before.
+            """
+            print(f"\t{items.title()} - {key[items]}")