+++ locations.py
@@ -1,4 +1,0 @@
-# Josh Lim
-# Comp Sci 30 P4
-# 11/07/2019
-# Dictionary and list containing all the locations