+++ player.py
@@ -5,6 +5,8 @@
 import numpy as np
 import math as m
 import tabulate as tab
+import map as mp
+import action as act
 
 
 class Player:
@@ -17,12 +19,10 @@ class Player:
         self.x = 1  # value for start tile
         self.y = 1  # value for start tile
 
-    def invGen(self):
-        """function for generating the inventory of player"""
-        invList = []
+    def viewInventory(self):
+        """function for viewing the inventory of the player"""
         for i in self.inventory:
-            print(f"\t{i}")  # prints list out
-            invList.append(i)  # adds each item to invList
+            print(f"\t{i}")  # prints list
 
     def description(self):
         """function for generating a description"""
@@ -48,3 +48,27 @@ class Player:
     def moveLeft(self):
         """functino for moving the player left"""
         self.x -= 1  # moves player to the left
+
+    def adjacentMoves(self):
+        """Returns all move actions for adjacent tiles."""
+        moves = []
+        if map.tileExists(self.x + 1, self.y):
+            moves.append(actions.moveRight)
+        if map.tileExists(self.x - 1, self.y):
+            moves.append(actions.moveLeft)
+        if map.tileExists(self.x, self.y - 1):
+            moves.append(actions.moveUp)
+        if map.tileExists(self.x, self.y + 1):
+            moves.append(actions.moveDown)
+        return moves
+
+    def availableActions(self):
+        """Returns all of the available actions in this room."""
+        moves = self.adjacentMoves()
+        moves.append(actions.viewInventory)
+        return moves
+
+    def doAction(self, action, **kwargs):
+        actionMethod = getattr(self, action.method)
+        if actionMethod:
+            actionMethod(**kwargs)