+++ CHANGELOG.md
@@ -38,9 +38,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  ### Added
  - Added movement.py file for the movement in the game
  - Added a xy based movement system in game.py
+ - Added classes to store map tile objects in
 
  ### Removed
-- Removed the old choose your own adventure like meny system
+- Removed the old choose your own adventure like menu system
 
  ### Fixed
  - Fixed a bug where rand center would duplicate center tile